Inline and prune
