+++ web/apps/photos/src/components/SearchBar.tsx
@@ -57,7 +57,7 @@ import {
     getDefaultOptions,
 } from "services/searchService";
 import { Collection } from "types/collection";
-import { SelectStyles } from "../styles/search";
+
 
 const { Option, ValueContainer, Menu } = components;
 
@@ -290,6 +290,72 @@ const SearchInputWrapper = styled(CenteredFlex, {
             }
         `}
 `;
+
+export const SelectStyles = {
+    container: (style) => ({
+        ...style,
+        flex: 1,
+    }),
+    control: (style, { isFocused }) => ({
+        ...style,
+        backgroundColor: "rgba(255, 255, 255, 0.1)",
+
+        borderColor: isFocused ? "#1dba54" : "transparent",
+        boxShadow: "none",
+        ":hover": {
+            borderColor: "#1dba54",
+            cursor: "text",
+        },
+    }),
+    input: (style) => ({
+        ...style,
+        color: "#fff",
+    }),
+    menu: (style) => ({
+        ...style,
+        marginTop: "1px",
+        backgroundColor: "#1b1b1b",
+    }),
+    option: (style, { isFocused }) => ({
+        ...style,
+        padding: 0,
+        backgroundColor: "transparent !important",
+        "& :hover": {
+            cursor: "pointer",
+        },
+        "& .main": {
+            backgroundColor: isFocused && "#202020",
+        },
+        "&:last-child .MuiDivider-root": {
+            display: "none",
+        },
+    }),
+    dropdownIndicator: (style) => ({
+        ...style,
+        display: "none",
+    }),
+    indicatorSeparator: (style) => ({
+        ...style,
+        display: "none",
+    }),
+    clearIndicator: (style) => ({
+        ...style,
+        display: "none",
+    }),
+    singleValue: (style) => ({
+        ...style,
+        backgroundColor: "transparent",
+        color: "#d1d1d1",
+        marginLeft: "36px",
+    }),
+    placeholder: (style) => ({
+        ...style,
+        color: "rgba(255, 255, 255, 0.7)",
+        wordSpacing: "2px",
+        whiteSpace: "nowrap",
+        marginLeft: "40px",
+    }),
+};
 
 const OptionWithInfo = (props) => (
     <Option {...props}>
